refactor: modal 공통 스타일 분리
