--- just-swim-app/src/app/_components/common/index.tsx
@@ -1,2 +1,0 @@
-export * from './calendar';
-export * from './imageCarousel';
