+++ just-swim-app/src/app/_components/common/modal/index.ts
@@ -1,0 +1,2 @@
+export * from './confirmModal';
+export * from './modalBody';
