+++ just-swim-app/src/app/_components/common/modal/modalBody/styles.module.scss
@@ -1,0 +1,44 @@
+@import '@/globals.scss';
+
+.modal_wrapper {
+  width: 100vw;
+  height: 100vh;
+
+  position: fixed;
+  
+  bottom: 0;
+  left: 0;
+
+  display: flex;
+  justify-content: center;
+
+  background-color: rgba(0, 0, 0, 0.4);
+
+  z-index: 200;
+
+  .modal {
+    width: 100%;
+
+    margin-top: auto;
+
+    background: #ffffff;
+
+    border-radius: 16px 16px 0 0;
+
+    overflow: hidden;
+
+    &_top_btn {
+      width: 100%;
+      height: 4px;
+      
+      padding-top: 10px;
+      
+      div {
+        background-color: $--gray-300;
+        width: 30px;
+        height: 4px;
+        margin: 0 auto;
+      }
+    }
+  }
+}
