+++ just-swim-app/src/app/_components/modal/colorModal/index.tsx
@@ -2,7 +2,7 @@
 
 import { MouseEvent, useState } from 'react';
 
-import { ColorPicker, ConfirmButton } from '@components';
+import { ColorPicker, ConfirmModal } from '@components';
 
 import styled from './styles.module.scss';
 
@@ -27,18 +27,12 @@ export function ColorModal({
     hideModal(event);
   }
 
-  const preventDefault = (event: MouseEvent<HTMLDivElement>) => {
-    event.preventDefault();
-  }
-
   return (
-    <div className={styled.modal_wrapper} onClick={preventDefault}>
+    <ConfirmModal 
+      hideModal={hideModal} 
+      confirmCallback={confirmSelectedColor}
+    >
       <div className={styled.modal}>
-        <button
-          className={styled.modal_top_btn}
-          onClick={hideModal}>
-          <div />
-        </button>
         <div className={styled.modal_title}>
           <p className="">구분 색</p>
           <p className="">스케쥴 정보 구분에 사용됩니다.</p>
@@ -49,11 +43,7 @@ export function ColorModal({
             change={changeSelectedColor}
           />
         </div>
-        <div className={styled.modal_btn}>
-          <ConfirmButton text='취소' kind='confirm-sub' onClick={hideModal} />
-          <ConfirmButton text='변경' kind='confirm' onClick={confirmSelectedColor} />
-        </div>
       </div>
-    </div>
+    </ConfirmModal>
   );
 }
