+++ just-swim-app/src/app/_components/modal/colorModal/styles.module.scss
@@ -1,85 +1,34 @@
 @import '@/globals.scss';
 
-.modal_wrapper {
-  width: 100vw;
-  height: 100vh;
+.modal {
+  width: 100%;
 
-  position: fixed;
-  
-  bottom: 0;
-  left: 0;
+  padding-top: 20px;
 
   display: flex;
-  justify-content: center;
+  flex-direction: column;
 
-  background-color: rgba(0, 0, 0, 0.4);
-
-  z-index: 200;
-
-  .modal {
-    width: 100%;
-    height: 337px;
-
-    margin-top: auto;
-
-    background: #ffffff;
-
-    border-radius: 16px 16px 0 0;
-
-    display: flex;
-    flex-direction: column;
-
-    overflow: hidden;
-
-    &_top_btn {
-      width: 100%;
-      height: 4px;
-      padding: 10px 165px;
-      
-      div {
-        background-color: $--gray-300;
-        width: 30px;
-        height: 4px;
-        margin: 0 auto;
-      }
+  .modal_title {
+    padding: 0 16px;
+  
+    p:nth-child(1) {
+      @include Title2-SB;
     }
-
-    &_title {
-      padding: 0 16px 28px;
-
-      p:nth-child(1) {
-        @include Title2-SB;
-      }
-
-      p:nth-child(2) {
-        @include Title5-M;
-        color: $--gray-600;
-      }
-    }
-
-    .color_list {
-      margin: 0 auto;
-      display: flex;
-      gap: 22px;
-      flex-wrap: wrap;
-      padding: 0 55px;
-    }
-
-    &_btn {
-      width: 100%;
-
-      margin: auto 0 20px;
-      padding: 0 16px;
-
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      gap: 10px;
-
-      & button:first-child {
-        background-color: $--Chips;
-      }
+  
+    p:nth-child(2) {
+      @include Title5-M;
+      color: $--gray-600;
     }
   }
+  
+  .color_list {
+    margin: 50px auto;
+    padding: 0 20px;
+
+    display: flex;
+    flex-wrap: wrap;
+
+    gap: 22px;
+  }
 }
+
