+++ just-swim-app/src/app/_components/modal/dateModal/index.tsx
@@ -2,7 +2,7 @@
 
 import { MouseEvent, useState } from 'react';
 
-import { DatePicker, ConfirmButton } from '@components';
+import { DatePicker, ConfirmModal } from '@components';
 
 import styled from './styles.module.scss';
 
@@ -27,27 +27,17 @@ export function DateModal({
     hideModal(event);
   }
 
-  const preventDefault = (event: MouseEvent<HTMLDivElement>) => {
-    event.preventDefault();
-  }
-
   return (
-    <div className={styled.modal_wrapper} onClick={preventDefault}>
+    <ConfirmModal 
+      hideModal={hideModal} 
+      confirmCallback={confirmSelectedDate}
+    >
       <div className={styled.modal}>
-        <button
-          className={styled.modal_top_btn}
-          onClick={hideModal}>
-          <div />
-        </button>
         <DatePicker
           selectedDate={initialDate}
           changeSelectedDate={changeSelectedDate}
         />
-        <div className={styled.modal_btn}>
-          <ConfirmButton text='취소' kind='confirm-sub' onClick={hideModal} />
-          <ConfirmButton text='변경' kind='confirm' onClick={confirmSelectedDate} />
-        </div>
       </div>
-    </div>
+    </ConfirmModal>
   );
 }
