+++ just-swim-app/src/app/_components/modal/dateModal/styles.module.scss
@@ -1,67 +1,12 @@
-@import '@/globals.scss';
-
-.modal_wrapper {
-  width: 100vw;
-  height: 100vh;
-
-  position: fixed;
-  
-  bottom: 0;
-  left: 0;
+.modal {
+  width: 100%;
 
   display: flex;
-  justify-content: center;
+  flex-direction: column;
 
-  background-color: rgba(0, 0, 0, 0.4);
+  padding-top: 50px;
 
-  z-index: 200;
-
-  .modal {
-    width: 100%;
-
-    margin-top: auto;
-
-    background: #ffffff;
-
-    border-radius: 16px 16px 0 0;
-
-    display: flex;
-    flex-direction: column;
-
-    overflow: hidden;
-
-    &_top_btn {
-      width: 100%;
-      height: 4px;
-      padding: 10px 165px;
-      
-      div {
-        background-color: $--gray-300;
-        width: 30px;
-        height: 4px;
-        margin: 0 auto;
-      }
-    }
-
-    .date_select {
-      padding: 20px 0;
-    }    
-
-    &_btn {
-      width: 100%;
-
-      margin: auto 0 20px;
-      padding: 0 16px;
-
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      gap: 10px;
-
-      & button:first-child {
-        background-color: $--Chips;
-      }
-    }
+  .date_select {
+    padding: 20px 0;
   }
 }
