+++ just-swim-app/src/app/_components/modal/dayModal/styles.module.scss
@@ -1,88 +1,33 @@
 @import '@/globals.scss';
 
-.modal_wrapper {
-  width: 100vw;
-  height: 100vh;
-
-  position: fixed;
-  
-  bottom: 0;
-  left: 0;
+.modal {
+  width: 100%;
 
   display: flex;
-  justify-content: center;
+  flex-direction: column;
 
-  background-color: rgba(0, 0, 0, 0.4);
-
-  z-index: 200;
-
-  .modal {
-    width: 100%;
-
-    margin-top: auto;
-
-    background: #ffffff;
-
-    border-radius: 16px 16px 0 0;
+  .day_list {
+    padding: 70px 0;
 
     display: flex;
-    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 20px;
 
-    overflow: hidden;
+    .day_button {
+      @include Title3-M;
 
-    &_top_btn {
-      width: 100%;
-      height: 4px;
-      padding: 10px 165px;
-      
-      div {
-        background-color: $--gray-300;
-        width: 30px;
-        height: 4px;
-        margin: 0 auto;
-      }
+      width: 50px;
+      height: 50px;
+
+      border-radius: 50%;
+
+      transition: all 0.2s;
     }
 
-    .day_list {
-      padding: 70px 0;
-
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 20px;
-
-      .day_button {
-        @include Title3-M;
-
-        width: 50px;
-        height: 50px;
-
-        border-radius: 50%;
-
-        transition: all 0.2s;
-      }
-
-      .day_button.selected {
-        background-color: $--Primary-blue;
-        color: white;
-      }
-    }
-
-    &_btn {
-      width: 100%;
-
-      margin: auto 0 20px;
-      padding: 0 16px;
-
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      gap: 10px;
-
-      & button:first-child {
-        background-color: $--Chips;
-      }
+    .day_button.selected {
+      background-color: $--Primary-blue;
+      color: white;
     }
   }
 }
