+++ just-swim-app/src/app/_components/modal/timeModal/index.tsx
@@ -2,7 +2,7 @@
 
 import { MouseEvent, useState } from 'react';
 
-import { ConfirmButton, TimePicker } from '@components';
+import { TimePicker, ConfirmModal } from '@components';
 
 import styled from './styles.module.scss';
 
@@ -26,27 +26,17 @@ export function TimeModal({
     hideModal(event);
   }
 
-  const preventDefault = (event: MouseEvent<HTMLDivElement>) => {
-    event.preventDefault();
-  }
-
   return (
-    <div className={styled.modal_wrapper} onClick={preventDefault}>
+    <ConfirmModal 
+      hideModal={hideModal}
+      confirmCallback={confirmSelectedTime}
+    >
       <div className={styled.modal}>
-        <button
-          className={styled.modal_top_btn}
-          onClick={hideModal}>
-          <div />
-        </button>
         <TimePicker
           value={selectedTime}
           setValue={changeSelectedTime}
         />
-        <div className={styled.modal_btn}>
-          <ConfirmButton text='취소' kind='normal' onClick={hideModal} />
-          <ConfirmButton text='변경' kind='confirm' onClick={confirmSelectedTime} />
-        </div>
       </div>
-    </div>
+    </ConfirmModal>
   );
 }
