+++ just-swim-app/src/app/_components/modal/timeModal/styles.module.scss
@@ -1,67 +1,14 @@
-@import '@/globals.scss';
-
-.modal_wrapper {
-  width: 100vw;
-  height: 100vh;
-
-  position: fixed;
-  
-  bottom: 0;
-  left: 0;
+.modal {
+  width: 100%;
 
   display: flex;
-  justify-content: center;
+  flex-direction: column;
 
-  background-color: rgba(0, 0, 0, 0.4);
-
-  z-index: 200;
-
-  .modal {
-    width: 100%;
-
-    margin-top: auto;
-
-    background: #ffffff;
-
-    border-radius: 16px 16px 0 0;
+  &_time_select {
+    padding: 50px 0;
 
     display: flex;
-    flex-direction: column;
-
-    overflow: hidden;
-
-    &_top_btn {
-      width: 100%;
-      height: 4px;
-      padding-top: 10px;
-      
-      div {
-        background-color: $--gray-300;
-        width: 30px;
-        height: 4px;
-        margin: 0 auto;
-      }
-    }
-
-    &_time_select {
-      padding: 50px 0;
-
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-
-    &_btn {
-      width: 100%;
-
-      margin: auto 0 20px;
-      padding: 0 16px;
-
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      gap: 10px;
-    }
+    align-items: center;
+    justify-content: center;
   }
 }
